+++ src/components/Errors.js
@@ -1,13 +1,20 @@
-
-import { Row, Col, Container, Jumbotron } from 'react-bootstrap';
+import { Row, Col, Container } from 'react-bootstrap';
 
 export const ErrorPage = ({ children }) =>
   <Container>
     <Row className="error-page">
       <Col>
-        <Jumbotron>
+        <div
+          // Bootstrap 5 legacy jumbotron 
+          style={{
+            padding: '2rem 1rem',
+            marginBottom: '2rem',
+            backgroundColor: '#e9ecef',
+            borderRadius: '.3rem',
+          }}
+        >
           {children}
-        </Jumbotron>
+        </div>
       </Col>
     </Row>
   </Container>;