+++ src/index.js
@@ -1,11 +1,10 @@
+export * from './components/CreateEditModal';
 export * from './components/DataTable';
 export * from './components/DragAndDropList';
-export * from './components/DataTable';
-export * from './components/DataTable';
+export * from './components/Errors';
 export * from './components/FormFields';
 export * from './components/IconButtons';
 export * from './components/LoadingIndicator';
-export * from './components/CreateEditModal';
 
 export * from './utils/Hooks';
 export * from './utils/TimeAndDate';